(Partial) Centrality Analysis.
Collected processing of the graph for analysis in a separate function.
In particular, edge weights are now normalised at the correct time,
i.e., as late as possible and not, say, *before* removing a node (the
Goalie) from the network.

Also cleaned up the notebook for analysing the centrality of each node
in regular situations vs power plays.  At the end, I perform a t-test
to see if the observed left-shift in centrality is significant.  The
tests say yes (yay); but, they are as of right now (arguably) invalid because PP networks from the same team and game are arguably not
independent.  (Arguably, so are regular networks from the same team
across different games, but there is not a lot of data as is.)

Then again, it is not clear when networks ever are independent because
the position networks generally aggregate multiple different players.

If only I would have had at least three statistics (-ish) courses so
far…